Moves gRPC client API annotations to grpc-api module.
