+++ microprofile/grpc/client/src/test/java/io/helidon/microprofile/grpc/client/EchoServiceTest.java
@@ -24,7 +24,9 @@ import java.util.concurrent.TimeoutException;
 import io.helidon.common.configurable.Resource;
 import io.helidon.common.tls.Tls;
 import io.helidon.grpc.api.Grpc;
+import io.helidon.grpc.api.GrpcChannel;
 import io.helidon.grpc.api.GrpcMarshaller;
+import io.helidon.grpc.api.GrpcProxy;
 import io.helidon.grpc.api.Unary;
 import io.helidon.microprofile.grpc.server.GrpcMpCdiExtension;
 import io.helidon.microprofile.testing.junit5.AddBean;
